Start adding a tooling CLI to hold all the logic for release builds.
It's a whole lot more straightforward to keep the decision trees and conditionals and other atomic operations in an actual programming language. Then the Github Actions workflows will just be responsible for setting up the environment and orchestrating invocations of the CLI.

For example, building and publishing the intl-message-database extension would now just be three pieces:
- `pnpm intl-cli db build --target ${TARGET}`
- `pnpm intl-cli db publish --target ${TARGET}`
- `pnpm intl-cli db publish-root`

All of the complexity around how the build actually happens on each platform and moving artifacts around is handled by the cli. That also means that local development can use the _same_ cli for building, making everything consistent and easy to navigate compared to different processes for local dev and CI.