testing out a real integration with react-native (#3)
There are inevitably changes that will have to be made to work with
react-native and Metro. One really annoying part of doing local dev with
that is that Metro is incapable of resolving things through symlinks,
and will isolate node_modules resolution in a really awkward way. So
this branch exists publicly to be able to sync changes into a host
project and test that way.