[bench] add a benchmark from icu-messageformat-parser to be able to compare (#4)
This copies the benchmark from
https://github.com/formatjs/formatjs/blob/f3bb540bea4614bbd22d4e80521d7a63101684e0/packages/icu-messageformat-parser/benchmark.js
to be able to compare roughly how well the ICUMarkdown parser performs.

Importantly, it will _never_ be as fast as the fastest pure-ICU parser,
since Markdown rules require a lot of extra logic that the parser has to
jump through. That said, hopefully it can still beat out the fastest
js-based parser (it doesn't currently...only about 60% as fast, but
still 4x faster than the old FormatJS parser).