[cli] Refactor commands to use a single base generator
Adding a simple package should now be as easy as adding a `createJsPackageCommand` call in `tools/index.js`, rather than having to make a whole folder and export a function with each individual thing.

This should also help with consistency across the board since they all create the same boilerplate.