[markdown][database] Support `select` and `selectordinal` icu placeholders
Turns out `select` is distinctly different from `plural` and `selectordinal`, so part of this is removing it from `IcuPluralKind` and handling all of the cases separately with a new struct. The rest is adding handling for `selectordinal` to the existing handling of `IcuPlural`.

In the CST, all of these types are handled distinctly, only in the AST are `plural` and `selectordinal` merged into one kind, matching the FormatJS structure with `pluralType`.