[markdown][database] Support "handler links" with no ICU syntax
This supports things like `[click me](onClick)`, where the target is actually a handler function and treated as an ICU variable, even though it doesn't have the ICU syntax on it.

This only works with single-token alphanumeric targets, so `'./target` doesn't count, `some_place` doesn't count, and any other link style doesn't count.

This is differentiated from ICU because the required type on the client side is almost always going to be different (`() => void` vs `string`).