[database] treat rich-text tags as "builtin", making them optional properties for all messages.
This isn't _entirely_ accurate since they still must be provided for every message, but saves some headache in the short term for dealing with type assignability.

In the future, this may switch to being two separate generic parameters so that assignability is more easily preserved _and_ enforced, since the bidirectionality is the current problem.