[runtime] constrain format functions with untyped `IntlMessageGetter` rather than specific typed instances.
When TypeScript evaluates a generic parameter on a union type, it tries to find a single common type for the parameter across all of the elements. When that constraint is a TypedIntlMessageGetter, the inner generic parameter then causes TypeScript to expect that all of the format values for each message are assignable to each other. If any message differs, it fails checking.

By extending the untyped `IntlMessageGetter` instead, TypeScript allows the entire union to be passed in as the type argument's value, meaning all of the downstream types are still able to check as expected, and union types of messages enforce typing for each element of the union, rather than just a common base type that might not be correct.