[markdown] a few very arbitrary performance improvements (~4%)
This adds what amounts to a pretty minor improvement to the lexer's performance by using lookup tables for some hotpath functions (utf8 char length for `advance` and significant byte detection in `consume_plain_token`). It also uses a `Cow<str>` in a few places during `process_cst_to_ast`, but the result isn't all that great.

By far the greatest amount time spent anywhere in the system is still `push_str` and `ToString` invocations, but i'm not entirely sure how to get rid of more of them. Moving ownership from the CST to the AST is important/awkward. Ideally it would be zero-copy as much as possible, but i don't know how to make that work with markdown's entity replacements and such