[markdown] 20% inline performance improvement by merging text tokens
Something I've known for a long time is that splitting tokens on every whitespace character adds _a lot_ of noise to the parsed structures for each message. Messages generally contain a lot of contiguous text and very little punctuation, but tokens were being split on whitespace, meaning every word was its own token with some piece of trailing trivia. That means _a lot_ of substrings get created, and there are even more `push_str` and other joining operations that happen downstream when processing into the CST and the AST.

This change merges contiguous text tokens with only whitespace between them into single tokens, since nothing within them can ever be significant in any other way to the parser. I expected some improvement from this, but on real world messages that are mostly just static text, this ends up yielding nearly 40% performance improvements, and still up to 25% improvements on messages where punctuation is common.

The only complicated part here is handling link destinations, which disallow whitespace in the destination to separate it from the optional title. So this also adds a new `LexContext` to handle that case and opt out of the text token merging.