[markdown] 3-5% performance improvement by using Rc<str>.
We'll assume that parsed messages never need to be sent across threads on the Rust side. If needed in the future, we can change this to Arc<str> instead, but I think it's unlikely, since the values are almost always ephemeral from the parser's creation until the resulting AST, which takes ownership over all values anyway and doesn't need the Rc.