[runtime] Restructure `IntlManager` to use composable formatters (#7)
This PR completely inverts the way `IntlManager` is structured to make
it more composable, extendable, and modular. Previously, `IntlManager`
defined formatting functions directly on itself, like `string` and
`formatToParts`, then subclasses of the class could add extra methods,
like `format` for React components. But this gets unwieldy for a variety
of reasons, namely that a single object is now responsible for
_defining_ multiple ways of formatting a message without a single
responsibility, and subclasses can't be composed to allow having
multiple subclassed formatting methods on the same manager.

This new structure still applies formatting functions as properties of
`IntlManager`, but they are created as bare functions _externally_ and
then applied to the manager with the `withFormatters` function, like so:

```typescript
import * as React from 'react';
import {
  astFormatter,
  IntlManager,
  makeReactFormatter,
  markdownFormatter, 
  stringFormatter
} from '@discord/intl';

const intl = new IntlManager('en-US').withFormatters({
  format: makeReactFormatter({
    $_: () => '',
    $i: (content: any) => <em>{content}</em>,
    $b: (content: any) => <strong>{content}</strong>,
    $p: (content: any) => <p>{content}</p>,
    $code: (content: any) => <code>{content}</code>,
    $link: ([target, ...content]: any[]) => {
      // Direct import from components/Anchor ensures circular dependencies through
      // the design barrel file don't cause "use before initialization" errors.
      const {Anchor} = require('@app/design/components/Anchor/web/Anchor');
      const props: React.ComponentProps<typeof Anchor> = {};

      if (typeof target === 'string') {
        props.href = SimpleMarkdown.sanitizeUrl(target)!;
      } else {
        props.onClick = target;
      }

      return <Anchor {...props}>{content}</Anchor>;
    },
  }),
  formatToPlainString: stringFormatter,
  formatToMarkdownString: markdownFormatter,
  formatToParts: astFormatter,
});

// Usage
intl.format(message, {}); // => React.ReactElement
intl.formatToMarkdownString(message, {}); // => string
```

A few things to note here:
- Some common formatting functions can be imported from the library,
including plain strings, markdown syntax, and a `RichTextNode` AST.
- `makeReactFormatter` allows consumers to supply their own rich text
elements to use within the formatter, but the implementation itself is
still abstracted away internally
- Usages still get called directly on the `intl` object, which is typed
to include the formatting functions as the return value from
`withFormatters`.

The implementation of this structure uses a new
`FormatterImplementation` type that defines a `format` function that
takes a message _instance_ and a set of values to return the final
result, and a new `FormatBuilder` type, which does the actual creation
of the result. This Builder type allows the final result to always be
created after a single pass over the AST, regardless of whether the
result is a plain string or an array of nodes or anything else. The
generic `StringBuilder` concatenates all values into a single string
value, while `ReactBuilder` and `AstBuilder` create a node tree from the
elements. Functionally, this is a big improvement over two-pass
implementation both in terms of complexity and performance, since the
total number of allocations can be greatly reduced, and iteration over
the tree only happens once.

The AST from `astFormatter` is also now a strictly-typed tree, rather
than a generic `Array<string | any>`, and it's structure is (roughly)
compatible with the structure from `SimpleMarkdown`, meaning it should
be a fairly straightforward drop-in replacement.

One of the biggest benefits of this change is migrating off of
FormatJS's internal `formatToParts` implementation and all of its AST
handling, meaning we are free to modify the generated and parsed AST
structures in any way we want for size and performance. Specifically,
that means `keyless-json` is now more viable and won't ever need to be
hydrated, saving a good 10% on space and 5-10% on loading time when an
application starts.