[runtime] export `loadAllMessagesInLocale`
This function provides a mechanism for an application to load _all_ messages from the application in the given locale and receive a single Promise to await it. Generally this isn't great since it adds an up-front cost to loading, but it is very friendly.

An alternative here is to just have modules act as their own boundaries and check their loader themselves using `someLoader.waitForLocaleLoaded()`.