[runtime] Use `string` for keys rather than `number`
Numbers are much easier to increment and pass along through a builder, but almost all React code in the wild shortcuts keys to be `string` rather than even `React.Key` most of the time, so putting `string` here is more friendly to that.

This also simplifies the FormatterReturnType` helper to just use `ReturnType<F>` rather than inferring from the type parameter of `FormatterImplementation`.