[runtime] add `reserialize` method for recreating original strings from asts.
This is useful when a message needs to be passed to another system that handles its own placeholder parsing. Note that reserialization is not _perfect_ because the original syntax is lossy when converting to a FormatJs-compatible AST, but it gets close enough that both markdown and placeholders are supported well enough.