[metro] fix extraImports config using absolute paths
Babel resolves import aliases before most plugins have a chance to work on the tree, meaning the original import paths with aliases like `@app/intl` are replaced with _relative_ paths, which are unusable for a querying perspective. This sets it up to resolve an absolute path according to the current source file, and then the consumer-side configuration can use `path.resolve(APP_FOLDER, 'relative/path')` as a key in the configuration object to get the same effect.