[runtime] add `runtimeHashMessageKey` to mirror message hashing on clients
Sometimes you need to take a string key and hash it to resolve a message at runtime on a client, but you can't currently because the hash function is inside a native Node extension. This _should_ be pretty rare, but it definitely happens.

Unfortunately, this means there have to be two implementations of the hashing, but that's why the naming is different (to avoid mistaking them unintentionally), and there are comments about keeping them in sync.