[runtime] Clarify temporary `ReactIntlMessage` type.
This is an unfortunate migration situation where we need to support both an old system where all messages are typed as `string` or `ReactNode` and this new system where React messages are _always_ an array of Elements. This type _should_ be compatible with both, but eventually should be replaced with the actual `ReactIntlMessage` from `formatters/react.ts`.