[database] add `getSourceFileMessageValues`
This method gets the current value of all messages for a given source file. If the source file is a definition, then the source translation is returned as the value for that message. If it is a translation, then the translation for that file's locale is returned instead. If no value exists for the key, or no translation exists for it, the key is still returned with a `None` value.