[loader-core] export index.d.ts types for typescript compat in consuming packages.
TypeScript often won't check js dependency files for types, meaning a consumer of this package won't actually get typed values for any of the exports.

Instead, there needs to be a `.d.ts` file that TypeScript detects and uses as the declared types for the whole package. After finding those, it'll happily traverse the js sources below it, but it has to start with a .d.ts still.