[database] add `exportTranslations`, separate `SourceFileMeta`.
`database.exportTranslations(extension)` is a new method on the database that will scan through the entire database, organizing all message values according to the file they should be persisted in based on the translations path from the source file meta and the locale of the message value. The `ExportTranslations` service does this and then directly writes the files back out to the system before returning the list of all file paths that were affected.

Consumers can use this result to check the that the files were written, inspect for empty files to prune, or remove any files that were not part of the export (since the entire database is exported, any file that exists but is not included must not be part of the project).

This method is distinct from the `precompile` method because it operates on the _current_ value of the database, not the list of messages that were originally sourced from a file. When performing imports of new translations, the database can be updated and then call `exportTranslations` to include those new values, even if there are new or removed keys, and the result will be accurate. Whereas with `precompile`, new translated keys would _not_ be added. This should probably be reconciled into a single method in the future, but for now this is the cleanest separation.