[markdown][database] Implement a true Visitor for message ASTs
Anything can now implement the `Visitor` trait and use `visit_with_mut` to walk the full message AST automatically, making implementations much easier to work with. `MessageVariablesVisitor` has been migrated to use this, and now it can also start building better validators for message content using it as well.

Ideally there would be some kind of `MultiVisitor` that allows multiple visitors to register for the same walk to help with cache locality when there are multiple to run, and also some way to control the traversal to skip things that are irrelevant to various visitors, but those can be built on top of this later on.