[runtime] more cautious error throwing when unsupported locales are loaded.
Throwing an error doesn't really make sense at runtime, since it's entirely likely that things won't always be translated in all locales. But if the configuration _says_ that a locale should be supported and it isn't actually present, then it's definitely an issue that should be resolved on the tooling level, not a bug in the consuming application, so it makes sense to throw and explain.