[runtime] Extract `intl-ast` as a separate package for managing the AST structure for messages
For one, it's just nice to have this separated for code hierarchy, but it also means it can more easily expanded to handle other format conversions. It also means that things like `intl_message_database/bench/native.js` can rely on the package rather than trying to copy over an implementation and have to keep it up to date.

This is in preparation for moving to KeylessJson by default.