[jest] Add jest-processor-discord-intl
This is a Jest processor that can compile definitions and translations suitably for running with Jest. It has slightly different behavior than the _loaders_ because it preserves the original source messages in the transpiled definition file (whereas the loaders create a separate asset file for smaller bundles).

Note that using this processor will still require using the `babel` plugin alongside it to transform consumer calls