[runtime] Allow multi-segmented literal strings in `intl.string()`
The current AST serialization method preserves special characters as their own entries in the result, such as `&`. For example, the message "Hello & Goodbye" would be serialized as:

```
["Hello ", "& ", "Goodbye"]
```

The old method of `intl.string(t.SOME_MESSAGE)` only checked that the first element of the AST array was a string, and then only _rendered_ that element as well. This change makes it so that it checks all entries in the array to be strings and then renders them all joined together.

Ideally the actual AST serialization would change here to save a bit more space, too, but this is close enough to start.