[runtime] Join plain text nodes for React formatting
This more closely matches FormatJS' behavior and expectations for things like Jest snapshots, where all contiguous text is combined into a single TextNode for the DOM. If they are kept as separate Nodes, snapshots tend to render them on multiple lines, which isn't really nice.

This is really just a compat change and not actually necessary, but avoids unnecessary drift when migrating to @discord/intl from FormatJs.