[types][runtime] Use empty object `{}` instead of `undefined` for plain messages
Originally it felt like using `undefined` was going to provide an easier way to discriminate between plain messages and those with formatting values in them, but it turns out it just causes _a lot_ of headaches, particularly around handling cases where both types of messages could be present. For example `TypedIntlMessageGetter<{count: any}> | TypedIntlMessageGetter<undefined>` doesn't do what you'd expect (either a message with a `count` variable or a message with no variables), it actually allows _any_ message in, because the way the typing is checked allows any object to be assignable to `undefined`.

Using `<{}>` looks a tiny bit uglier and isn't as obvious at a glance, but it makes a lot of the typing logic simpler and avoids the need for unions in most places. Most cases are handled just by switching the order from `T extends undefined` to `{} extends T` to preserve the same "T has no values" kind of semantic.

Eventually I think I'd like to make a public type alias for a `PlainMessageGetter`, but I think that's also more headache than it's worth.