[database][loader][bundler] Handle `secret` and `translate` metas properly
The `secret` meta property for messages and source files is now actually respected when _bundling_ messages (renamed from `precompile` internally, but the public loader method is still called `precompileFileForLocale`. When a message is `secret`, it will default to being obfuscated in the bundled output for a file, but if `bundleSecrets` is set to true in the compilation options, the message is preserved (useful for testing in development but hiding in production, for example).

`bundleSecrets` has also been removed as a meta property on sources and is instead now a configuration option passed into loaders to determine how messages should be handled on the fly. In that same vein, configuration options for `precompile` have been converted to an object to make adding more options easier in the future.