[database][metro] Unfortunate concessions for metro multiprocessing
As noted in the comment, Metro's multiprocessing model means we can't trust a message to be defined when bundling, since it may have happened in a separate process (and thus a different instance of the database).

Ideally we'd have some mechanism to ensure the database is a pure singleton, possibly running in its own process entirely for Metro, but that doesn't exist yet and would be pretty bad for read-oriented performance.

For now, we'll just go back to bundling messages even if they have no definition.