[markdown][ast][runtime] Refactor Links to use new `control` field.
`control` is not part of FormatJS and may break expected behavior. It's added as a way of discriminating between two kinds of children: "visible" and "controlling". For example, links have a visible component (the label) and a control component (the destination). Compiling these into a single `children` list becomes extremely difficult to process effectively when rendering because it can only rely on convention to establish what "visible" and what is "hidden". By adding this `control` type, the distinction is always trivial to process.

This also makes it easier to distinguish between the two for cases where rendering only cares about visible elements. For example `intl.string` only renders the _label_ of a link and should completely ignore the controlling portion. This was already doable just by ignoring the first child, but now `children` can be copied wholesale with no worry about what it contains. This will likely enable more in the future, too.

This also means we can get rid of the `$_` empty placeholder, which is nice in its own right.