[validator] Validate repeated plural options
Plural selectors are "required" to have unique arms. At the very least, it's objectively incorrect to have more than one option with the same name, since only one can ever match. For example:

```
{count, plural, one {# object} other {# objects} other {# things}}
```

Here, `other` is repeated, and when the renderer comes to match it, it could pick either instance depending on how it was implemented! FormatJS I think throws an error or something when it encounters this, but since we're explicitly compiling and linting, we can just make it a lint and let the fallback happen the rest of the time. The consumer can decide how to manage the lint severity.