[eslint] Perform and report native validations on definitions.
The `native-validation` rule runs the native database's validation rules once per file and then reports all of the diagnostics it finds coming from that file.

This is still repetitive and should probably be deduplicated somehow, but it needs to be sure that it has all of the validations for any new files that have been added to the database since the last time it ran.

That said, this could/should probably be split into multiple rules on the eslint side, so that the severity can be controlled by the consumer rather than all being bundled into one rule that always errors or warns regardless of the diagnostic.