[runtime] avoid `.reduce()` and `.bind` when creating message binds
`.reduce()` is definitely slow on an object with a vast number of keys, and using a direct iteration should be quite a bit faster. Also, the `Object.keys` call is meaningless, since `messageKeys` is already an array, not an object (actually, it _was_, but only by accident. The transformer creating the loader should do this work ahead of time).

I also have a feeling that using `.bind()` is taking more work than creating an arrow function directly inline, so I want to see what kind of performance difference that makes.

In our app, the `getBinds` function is currently taking ~200ms on low end android phones, so my hope is that this will reduce that to maybe ~60ms.