[eslint][validator] Refactor native validations to one-rule-per-diagnostic
This gives better control for consumers to turn on and off different native validations as desired. All rules are enabled in the `recommended` set, but now the individual native rules can be split up. This tries to be efficient about _processing_ them, but the navigation is definitely not the best, since it runs a lot of duplicate work determining if a node is a definition or not.

Some memoization there is definitely doable.