[loader] Experiment with creating message binds at build time rather than at runtime.
When a messages file contains multiple thousands of keys (for example, our largest contains ~18k), iterating over all of the messages in a single go to generate bound functions is a very non-negligible amount of time where the engine _cannot_ do anything else. This only really matters on low-end devices, but in hopes of mitigating some of that, this moves the bind generation to "bundle time" (sort of) by doing codegen for each bind directly rather than having it happen through the `getBinds` method at runtime.

This is not ideal, but in testing locally, it seems like it might give back a decent chunk of time.