[loader][metro][rspack] Allow switching between `preGenerateBinds`.
`preGenerateBinds` has shown to speed up initialization in Hermes apps by 2-3x. For most phones, this is entirely trivial and meaningless, but for older low-end phones this can be ~100-150ms of time for a surprisingly small trade on space, since Hermes compiles to bytecode rather than preserving source code.

As such, this is now an option on `MessageDefinitionsTransformer` and defaults to `true` for metro builds.