[runtime] Return loading promise from `_loadLocale` when it already exists.
This caused things like `waitForAllDefaultIntlMessagesLoaded` and `loadAllMessagesInLocale` to return true _immediately_ if the locale was _already_ in progress loading, because it wouldn't return the same loading promise, but rather a new one that resolves as soon as the function returns.

Now they should properly wait. To ensure this still escapes early when the locale is already loaded, the existence check has been moved to the top as well.