Safer secrets through pre-loading all messages (#12)
`secret` messages are an important feature of this system, allowing
developers to mark a message or set of messages as "hidden from
clients". When this flag is set on either a message or source file meta,
the message will only be bundled as a non-reversible hash, ensuring that
features can be developed in _secret_ without leaking anything about
them to users (at least, not through plain text messages).

For Discord, this lets teams work entirely in-band, while allowing
messages to still go through the normal translation flows and be ready
to go when the team is ready to launch their feature, replacing a bunch
of previous manual workarounds involving long-lived feature branches,
copy-pasting json for messages to get translated, holding off on merges,
and manually obfuscating text for release builds. Now, the entire
process is automated and handled just like normal messages, all
controlled by a single boolean flag in the source file.

One important caveat with this feature is that _because translation
files can be merged together and shared between definitions files, the
database needs to already know _all_ message definitions before it can
accurately bundle translations for the final build. Otherwise, a
translation with no "known" definition at the time of bundling will
either get omitted entirely (safer for secrets, but will cause crashes
when expected messages from another source file are missing), or get
bundled as plain text (because the definition marking the message as
`secret` has not yet been encountered). To solve this, this PR also adds
new `findAllMessagesFiles` and `processAllMessagesFiles` methods to the
database and loader that will walk the entire project directory (or list
of specified `watchFolders`) to discover all message definitions ahead
of time. The Metro and Rspack loaders then use this preloaded database
to accurately bundle translations as they are encountered. It's an
unfortunate step that is thankfully not too costly (since the walking
and processing can happen in across multiple threads).