[rspack] don't allow translations to be cached
Because translations files can change based on the source file changing (i.e., the definition changes `secret` on a message, requiring the translations to also make that value secret), it's not possible/feasible to cache the result of loading translations, since they won't get properly invalidated.

Ideally, this would be done automatically with the `addDependency` feature of webpack/rspack, but there's not a great way to add a "reverse" dependency (i.e., tell rspack that another file is dependent on this file's changes). That's possible in a plugin, though, so maybe it's more impetus to convert from a loader to a plugin soon.