fix `reserialize()` of `InternalIntlMessage` (#13)
This fixes `serializeAst()` and `serializeAstTag()` used by
`InternalIntlMessage.reserialize()`.

I've tested the function on intl messages used by the Discord client and
compared with their versions from the old system. Most of them are the
exact same, with the rest being only slightly different (aka it works
well!).

The only thing I'm unsure about is the node index in `$link` handler.
The only enum values with index 3 are `Offset` and `Control`, none of
which seem to fit here so I left it hardcoded.