[loader] Fix watch mode working as a one-shot
CI machines seem to not like processing file events fast enough for them to run before the function exits? Or at least somehow the compiled assets aren't visible to metro's graph from builds on CI despite working. I think this might also be due to a limited number of threads to process the events, but it's too hard to say really.