[metro] Use a single project-wide cache dir for compiled translations
Rather than storing a `.cache/intl` folder within every module that contains translations, the asset plugin will now resolve assets to a single cache directory at the root of the project. The config loader also asserts that different parts of the config for that path are valid and will function as expected.