[intl] Fix rendering objects using `AstBuilder`
the `object` type node wasn't really a proper thing to exist. When hooks and other tags return an object, it's typically not a custom type, but rather just an externally-created node that should be as valid as any other known node. So now `AstBuilder` will directly append the object rather than wrapping it in a separate node type.

This also adjusts the typing to make that a little bit more sound, both for AstBuilder and ReactBuilder, although clients don't really see these types.