[exporter] Explicitly sort keys when bundling
For the sake of diff stability, we want to ensure that all precompiled files have a consistent ordering of values. I had assumed that this would happen by default since the hashing algorithm is consistent, but it turns out it can (and does) vary wildly between runs.

To fix that, this adds an intermediate vector in the IntlMessageBundler service that collects all of the message keys that are going to be written and sorts the list before iterating to write them to the target buffer. Using the unstable `collect_into` feature, this makes it possible to reserve a vector with enough capacity ahead of time, making the insertion much more efficient. So much more efficient, that somehow this is actually _faster_ than the previous iteration over the Set was, likely because it becomes a linear iteration over u64s rather than having potential dereferences for each item in the Set.