[metro] Fix paths for Metro asset plugin in local development
Upgrading from Metro 0.80.x to 0.81.0 seems to have broken asset loading when the httpServerLocation doesn't effectively match the path for assets on disk. This change defaults the asset plugin to writing matching paths for the http and file system locations, allowing translations to be loaded during development again.

This can be configured with the `publicPath` property both on Metro's transformer config and the plugin config itself to test out more changes in the future.