[runtime] add `string` to `ReactIntlMessage` alias.
`ReactIntlMessage` should represent "any message-like content", mostly just meaning that it's not arbitrary react nodes being passed around. But we also want to allow `string` for the case of plain-string messages _and_ for users to have a single type that still allows basic values to be passed in (like formatting with `intl.string`).