[loader][rspack][metro] Enable `proxy` mode for pregenerating binds.
This adds a new value for `pregenerateBinds` with `'proxy'`, which will tell the transformer to create the `binds` object for a messages file using a `Proxy`. This can potentially skip thousands of function bind instantiations during startup, which should help large message files resolve faster, especially on mobile with Hermes.

The proxy also implements `has` to ensure queries like `key in messages` still function as expected (used in `MessageLoader`), and once a bind is generated it replaces the Proxy trap with the loader value directly, meaning repeated accesses have zero overhead.