[loader] Try skipping keys entirely for proxy bind mode. (#19)
This is a little dangerous, since the proxy object will _always_ return
a message getter, even if the underlying data doesn't actually contain
the message, but in general that's not an issue since typescript and
linting rules should ensure that no completely-unknown messages are
found. And even in that case, `MessageLoader` is set up in such a way
that an undefined message value throws an error regardless. This just
moves that kind of error a little further down the road, but also ends
up giving more context to it (describing which locale is missing a
value, rather than just saying `<key> is undefined`).