[runtime][loader] Move Proxy generation to runtime package and enable chaining through a single method.
The `chainMessagesObjects` method should work with both literals and proxies interchangeably, meaning any consuming code can call it with two arbitrary messages objects and get a combined result back.

This, in general, is not an ideal solution, but is generally helpful for us, where we have two old systems that used to be combined at runtime and we want to preserve that behavior until we're able to migrate more messages into modular objects.