[allocator] Extract allocator crate to be able to build packages independently.
Without this, inevitably `#[global_allocator]` ends up appearing twice in a built crate, which can't be compiled. Having it be external means there will only ever be one instance of it, and crates can opt-in to it when they build.

Adding this to the JSON parser shaves ~12% off the runtime of the parser. I'm pretty confident that the majority of the time spent in it is allocating the iterable array and converting into JS string objects, though.