[types] Make spurious variables required
Spurious variables being optional was a temporary placeholder as we migrated systems, since there were a large number of callsites that weren't supplying all of the necessary variables and would fail type checks. Now that we've fully migrated, we can start enforcing this for correctness and to avoid runtime crashes.