[json] Create js array directly from message iterator
This bypasses the intermediate `Vec<Message>` that was created solely to pass on to napi-rs' internal conversion to an `Array`. However, there's no real performance gain because the incremental insertion still means the array has to get resized multiple times before it's completed, and we can't give much more of an accurate size hint since the original parse is a stream as well.