[core][exporter] Canonicalize translations paths (fixes #22)
From the description, I had assumed that `path::absolute` resolved `..` segments in paths to return a canonicalized path as a result, but it turns out that `..` is actually a symlink, and `absolute` says it does not handle symlinks, so that it doesn't have to talk to the filesystem.

So this changes the `get_translations_path` to actually return the canonical path and also return a proper `Err` if the file can't be found.