[core] Fix ENOENT error when translations path doesn't exist
`fs::canonicalize` will return an `Err` if the path doesn't exist, which I tried to work around by not adding the `locale` to the path until after canonicalization, but it's still possible that the entire directory doesn't exist, or any path to it. So instead of doing that, lets use a crate that can just normalize the path without talking to the file system and then use `fs::create_dir_all` to ensure it exists before writing.